스모킹 세이브 리턴값 실시간 흡연량으로 변경

+++ gbt/src/main/java/ksafinalproject/gbt/smoking/controller/SmokingController.java
@@ -25,12 +25,12 @@ public class SmokingController {
     private final SmokingService smokingService;
 
     @PostMapping("")
-    public int smokingSave(@RequestBody ISmoking iSmoking) {
+    public Long smokingSave(@RequestBody ISmoking iSmoking) {
         try {
             return smokingService.saveSmoking(iSmoking);
         } catch (Exception e) {
             log.error("Error : {}", e.getMessage());
-            return -1;
+            return -1L;
         }
     }
